+++ packages/building-rollup/demo/mpa/js/subpage-dep.js
@@ -1,0 +1,1 @@
+export const subpageDepMetaUrl = import.meta.url;