+++ packages/building-rollup/src/types.d.ts
@@ -1,3 +1,5 @@
+import { PluginOptions } from '@open-wc/rollup-plugin-html';
+
 export interface BasicOptions {
   outputDir?: string;
   nodeResolve?: boolean | object;
@@ -5,15 +7,19 @@ export interface BasicOptions {
   terser?: boolean | object;
   legacyBuild?: boolean;
   developmentMode?: boolean;
+  rootDir?: string;
 }
 
 export interface SpaOptions extends BasicOptions {
-  html?: boolean | object;
+  html?: boolean | PluginOptions;
   polyfillsLoader?: boolean | object;
   workbox?: boolean | WorkboxOptions;
   injectServiceWorker?: boolean;
 }
 
+export interface MpaOptions extends SpaOptions {}
+
 interface WorkboxOptions {
-  swDest: string
+  swDest?: string;
+  globDirectory?: string;
 }