+++ packages/dedupe-mixin/src/dedupeMixin.js
@@ -1,28 +1,30 @@
 const appliedClassMixins = new WeakMap();
 
-function getPrototypeChain(obj) {
-  const chain = [];
-  let proto = obj;
-  while (proto) {
-    chain.push(proto);
-    proto = Object.getPrototypeOf(proto);
-  }
-  return chain;
-}
-
-function wasApplied(mixin, superClass) {
-  const classes = getPrototypeChain(superClass);
-  for (const klass of classes) {
+/** Vefify if the Mixin was previously applyed
+ * @private
+ * @param {function} mixin      Mixin being applyed
+ * @param {object} superClass   Class receiving the new mixin
+ * @returns {boolean}
+ */
+function wasMixinPreviouslyApplied(mixin, superClass) {
+  let klass = superClass;
+  while (klass) {
     if (appliedClassMixins.get(klass) === mixin) {
       return true;
     }
+    klass = Object.getPrototypeOf(klass);
   }
   return false;
 }
 
+/** Apply each mixin in the chain to make sure they are not applied more than once to the final class.
+ * @export
+ * @param {function} mixin      Mixin to be applyed
+ * @returns {object}            Mixed class with mixin applied
+ */
 export function dedupeMixin(mixin) {
   return superClass => {
-    if (wasApplied(mixin, superClass)) {
+    if (wasMixinPreviouslyApplied(mixin, superClass)) {
       return superClass;
     }
     const mixedClass = mixin(superClass);