+++ packages/demoing-storybook/CHANGELOG.md
@@ -3,6 +3,62 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.3.18](https://github.com/open-wc/open-wc/compare/@open-wc/demoing-storybook@2.3.17...@open-wc/demoing-storybook@2.3.18) (2020-08-07)
+
+**Note:** Version bump only for package @open-wc/demoing-storybook
+
+
+
+
+
+## [2.3.17](https://github.com/open-wc/open-wc/compare/@open-wc/demoing-storybook@2.3.16...@open-wc/demoing-storybook@2.3.17) (2020-08-05)
+
+**Note:** Version bump only for package @open-wc/demoing-storybook
+
+
+
+
+
+## [2.3.16](https://github.com/open-wc/open-wc/compare/@open-wc/demoing-storybook@2.3.15...@open-wc/demoing-storybook@2.3.16) (2020-08-04)
+
+**Note:** Version bump only for package @open-wc/demoing-storybook
+
+
+
+
+
+## [2.3.15](https://github.com/open-wc/open-wc/compare/@open-wc/demoing-storybook@2.3.14...@open-wc/demoing-storybook@2.3.15) (2020-08-03)
+
+**Note:** Version bump only for package @open-wc/demoing-storybook
+
+
+
+
+
+## [2.3.14](https://github.com/open-wc/open-wc/compare/@open-wc/demoing-storybook@2.3.13...@open-wc/demoing-storybook@2.3.14) (2020-07-29)
+
+**Note:** Version bump only for package @open-wc/demoing-storybook
+
+
+
+
+
+## [2.3.13](https://github.com/open-wc/open-wc/compare/@open-wc/demoing-storybook@2.3.12...@open-wc/demoing-storybook@2.3.13) (2020-07-24)
+
+**Note:** Version bump only for package @open-wc/demoing-storybook
+
+
+
+
+
+## [2.3.12](https://github.com/open-wc/open-wc/compare/@open-wc/demoing-storybook@2.3.11...@open-wc/demoing-storybook@2.3.12) (2020-07-10)
+
+**Note:** Version bump only for package @open-wc/demoing-storybook
+
+
+
+
+
 ## [2.3.11](https://github.com/open-wc/open-wc/compare/@open-wc/demoing-storybook@2.3.10...@open-wc/demoing-storybook@2.3.11) (2020-06-25)
 
 **Note:** Version bump only for package @open-wc/demoing-storybook