+++ packages/karma-esm/CHANGELOG.md
@@ -3,6 +3,65 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [3.0.3](https://github.com/open-wc/open-wc/compare/@open-wc/karma-esm@3.0.2...@open-wc/karma-esm@3.0.3) (2020-08-10)
+
+**Note:** Version bump only for package @open-wc/karma-esm
+
+
+
+
+
+## [3.0.2](https://github.com/open-wc/open-wc/compare/@open-wc/karma-esm@3.0.1...@open-wc/karma-esm@3.0.2) (2020-08-04)
+
+**Note:** Version bump only for package @open-wc/karma-esm
+
+
+
+
+
+## [3.0.1](https://github.com/open-wc/open-wc/compare/@open-wc/karma-esm@3.0.0...@open-wc/karma-esm@3.0.1) (2020-07-29)
+
+**Note:** Version bump only for package @open-wc/karma-esm
+
+
+
+
+
+# [3.0.0](https://github.com/open-wc/open-wc/compare/@open-wc/karma-esm@2.16.18...@open-wc/karma-esm@3.0.0) (2020-07-24)
+
+
+### Features
+
+* **testing-karma:** use up to date coverage reporter ([e853364](https://github.com/open-wc/open-wc/commit/e853364b9dde1924199f4309cf6689b3bcc6e25f)), closes [#1164](https://github.com/open-wc/open-wc/issues/1164)
+
+
+### BREAKING CHANGES
+
+* **testing-karma:** coverageIstanbulReporter object in karma config has been replaced with coverageReporter
+move from karma-coverage-istanbul-reporter to karma-coverage
+
+See MIGRATION.md for more info.
+
+
+
+
+
+## [2.16.18](https://github.com/open-wc/open-wc/compare/@open-wc/karma-esm@2.16.17...@open-wc/karma-esm@2.16.18) (2020-07-24)
+
+**Note:** Version bump only for package @open-wc/karma-esm
+
+
+
+
+
+## [2.16.17](https://github.com/open-wc/open-wc/compare/@open-wc/karma-esm@2.16.16...@open-wc/karma-esm@2.16.17) (2020-07-10)
+
+**Note:** Version bump only for package @open-wc/karma-esm
+
+
+
+
+
 ## [2.16.16](https://github.com/open-wc/open-wc/compare/@open-wc/karma-esm@2.16.15...@open-wc/karma-esm@2.16.16) (2020-06-25)
 
 **Note:** Version bump only for package @open-wc/karma-esm